+++ frontend/src/routes/recipes/+page.svelte
@@ -1,29 +1,37 @@
 <script lang="ts">
-	import 'iconify-icon'
+	import 'iconify-icon';
 	import { goto } from '$app/navigation';
 	import { recipes, filterTerm, user } from '$lib/store';
 	import RecipeCard from './RecipeCard.svelte';
-	import { getToastStore, type ToastSettings } from '@skeletonlabs/skeleton';
+	import { getToastStore } from '@skeletonlabs/skeleton';
+	import { sendToast } from '$lib/utils';
 	// TODO: hide the recipes that are 'mix' ingredients such as 'simple syrup'
 
 	const toastStore = getToastStore();
 
 	function handleKeyPress(event: KeyboardEvent) {
 		if (event.key == 'Enter') {
-			handleClick();
+			// TODO: open drawer only if there are no matching ingredients
+			const list = $recipes.filter((recipe) =>
+				recipe.name.toUpperCase().includes($filterTerm.toUpperCase())
+			);
+			if (list.length === 0) {
+				handleClick();
+				return;
+			} else {
+				const inputElement = document.getElementById('recipe-search');
+				inputElement?.blur();
+			}
 		}
 	}
 
 	function handleClick() {
 		if (!$user) {
-			const t: ToastSettings = {
-				message: 'You must be logged in',
-				background: 'variant-filled-error'
-			};
-			toastStore.trigger(t);
+			const message = 'You must be logged in to add recipes.'
+			sendToast(toastStore, message)
 			return;
 		}
-		goto("/recipes/new")
+		goto('/recipes/new');
 	}
 
 	function clearInput() {
@@ -41,6 +49,7 @@
 		<!-- TODO: exit the search bar on enter press -->
 		<div class="w-72 input-group grid-cols-[240px_auto]">
 			<input
+				id="recipe-search"
 				type="search"
 				placeholder="Search {$recipes.length} ingredients..."
 				bind:value={$filterTerm}
@@ -55,9 +64,8 @@
 		<ul class="space-y-2">
 			{#each $recipes as recipe}
 				<RecipeCard {recipe} />
-
 			{/each}
 		</ul>
-        <button on:click={handleClick} class="btn variant-filled-success w-full">Add new recipe</button>
+		<button on:click={handleClick} class="btn variant-filled-success w-full">Add new recipe</button>
 	</div>
 </div>