+++ frontend/tailwind.config.ts
@@ -9,7 +9,11 @@ export default {
 	darkMode: 'class',
 	content: ['./src/**/*.{html,js,svelte,ts}', join(require.resolve('@skeletonlabs/skeleton'), '../**/*.{html,js,svelte,ts}')],
 	theme: {
-		extend: {},
+		extend: {
+			screens: {
+				'tall': { 'raw': '(min-height: 600px)' },
+			  }
+		},
 	},
 	plugins: [
 		forms,