+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.1
+
+- feat: clear button on filter search
+- feat: remove features when user is not logged in
+- feat: make ingredient label responsive for checkbox
+- feat: UX improvement on ingredients/new
+- feat: improve UX for 'new' button when user is not logged in
+
 ## 0.1.0
 
 - register/login/logout/update: users